Adds refined type validation for query params
It's not ideal though since I had to hack around to get human-friendly messages,
the messages are not going to compose (e.g. email being non-empty and containing
an '@' would need a specific error message, without re-using the one of non-empty)
and it's probably going to require manual work for entity validation (i.e.
custom validation of each type)